QA ministack setup

+++ le-docker/src/main/scripts/nodejs/build.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
-docker rmi -f $(docker images -a --filter "dangling=true" -q --no-trunc) 2> /dev/null || true
+
+DKRMI_FLAG="${DKRMI_FLAG:=true}"
+($DKRMI_FLAG) && docker rmi -f $(docker images -a --filter "dangling=true" -q --no-trunc) 2> /dev/null || true
 
 source ../functions.sh
 
@@ -7,4 +9,4 @@ pushd le-express
 build_docker latticeengines/express
 popd
 
-docker rmi -f $(docker images -a --filter "dangling=true" -q --no-trunc) 2> /dev/null || true
+($DKRMI_FLAG) && docker rmi -f $(docker images -a --filter "dangling=true" -q --no-trunc) 2> /dev/null || true
